Fix a bug that stops initial value from being set
The Preference superclass never called onSetInitialValue() because
onGetDefaultValue() was not defined in ColorPickerPreference.
onGetDefaultValue() is now defined to fix the issue.

Additionally, ColorPickerPreference.mDefaultValue has been removed since
it is redundant with Preference.mDefaultValue. Strings containing hex
values in the form of `#000000` will no longer be supported as a
defaultValue. However, traditional hex values in the form of `0x000000`
can still be used when defining a default Integer value in XML.

SeekBarPreference (an integer-based preference written by the Android
developers) was consulted when making these changes.