Fix for setting DefaultValue as hex string
I have noticed that when you set defaultValue you can only set it as an actual color integer.

1. If you set it as hex string it crashes, as it expects integer
2. If you set it as reference to resources color, it persists 0 into shared preferences

Method onGetDefaultValue was modified to look for a string and if it is not null, to use built in method to parse hex string into color integer and return it as default. Color hex can only be used with 6 or 8 chars. If the string is null, it will look for an integer and use it.